// src/caseStudies/types.ts
export interface CaseStudyData {
  title: string
  category: string
  description: string
  result: string
  challenge: string
  solution: string
  impact: string
  process: string
  tools: string[]
  timeline: string
  outcome: string
  image: string
  heroImage?: string
  projectImage1?: string
  projectImage2?: string
  projectImage3?: string
  projectImage4?: string
}